+++ extensions/zed-recent-projects/CHANGELOG.md
@@ -1,5 +1,9 @@
 #  Zed Recent Projects Changelog
 
+## [Fixes] - {PR_MERGE_DATE}
+
+- Handle Zed sqlite schema version 27
+
 ## [Fixes] - 2025-09-08
 
 - Use latest Zed sqlite schema